# backend/MNTRK_API/swagger_server/models/detection_pattern_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.habitat_analysis_request_geographic_coordinates import HabitatAnalysisRequestGeographicCoordinates  # noqa: F401,E501
from swagger_server import util


class DetectionPatternResponse(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: Object=None, location: HabitatAnalysisRequestGeographicCoordinates=None):  # noqa: E501
        """DetectionPatternResponse - a model defined in Swagger

        :param id: The id of this DetectionPatternResponse.  # noqa: E501
        :type id: Object
        :param location: The location of this DetectionPatternResponse.  # noqa: E501
        :type location: HabitatAnalysisRequestGeographicCoordinates
        """
        self.swagger_types = {
            'id': Object,
            'location': HabitatAnalysisRequestGeographicCoordinates
        }

        self.attribute_map = {
            'id': 'id',
            'location': 'location'
        }
        self._id = id
        self._location = location

    @classmethod
    def from_dict(cls, dikt) -> 'DetectionPatternResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The DetectionPatternResponse of this DetectionPatternResponse.  # noqa: E501
        :rtype: DetectionPatternResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> Object:
        """Gets the id of this DetectionPatternResponse.


        :return: The id of this DetectionPatternResponse.
        :rtype: Object
        """
        return self._id

    @id.setter
    def id(self, id: Object):
        """Sets the id of this DetectionPatternResponse.


        :param id: The id of this DetectionPatternResponse.
        :type id: Object
        """

        self._id = id

    @property
    def location(self) -> HabitatAnalysisRequestGeographicCoordinates:
        """Gets the location of this DetectionPatternResponse.


        :return: The location of this DetectionPatternResponse.
        :rtype: HabitatAnalysisRequestGeographicCoordinates
        """
        return self._location

    @location.setter
    def location(self, location: HabitatAnalysisRequestGeographicCoordinates):
        """Sets the location of this DetectionPatternResponse.


        :param location: The location of this DetectionPatternResponse.
        :type location: HabitatAnalysisRequestGeographicCoordinates
        """

        self._location = location