# backend/MNTRK_Agent_API/swagger_server/controllers/default_controller.py
import connexion
import six

from swagger_server.models.community_observation_request import CommunityObservationRequest  # noqa: E501
from swagger_server.models.community_observation_response import CommunityObservationResponse  # noqa: E501
from swagger_server.models.data_management_open_request import DataManagementOpenRequest  # noqa: E501
from swagger_server.models.data_management_open_response import DataManagementOpenResponse  # noqa: E501
from swagger_server.models.data_management_transform_request import DataManagementTransformRequest  # noqa: E501
from swagger_server.models.data_management_transform_response import DataManagementTransformResponse  # noqa: E501
from swagger_server.models.detection_pattern import DetectionPattern  # noqa: E501
from swagger_server.models.detection_pattern_response import DetectionPatternResponse  # noqa: E501
from swagger_server.models.explain_request import ExplainRequest  # noqa: E501
from swagger_server.models.explain_response import ExplainResponse  # noqa: E501
from swagger_server.models.geospatial_analysis_request import GeospatialAnalysisRequest  # noqa: E501
from swagger_server.models.geospatial_analysis_response import GeospatialAnalysisResponse  # noqa: E501
from swagger_server.models.habitat_analysis_request import HabitatAnalysisRequest  # noqa: E501
from swagger_server.models.habitat_prediction import HabitatPrediction  # noqa: E501
from swagger_server.models.io_t_ingest_response import IoTIngestResponse  # noqa: E501
from swagger_server.models.model_training_request import ModelTrainingRequest  # noqa: E501
from swagger_server.models.model_training_response import ModelTrainingResponse  # noqa: E501
from swagger_server.models.rag_query_request import RAGQueryRequest  # noqa: E501
from swagger_server.models.rag_query_response import RAGQueryResponse  # noqa: E501
from swagger_server.models.risk_analysis_request import RiskAnalysisRequest  # noqa: E501
from swagger_server.models.risk_analysis_response import RiskAnalysisResponse  # noqa: E501
from swagger_server.models.video_stream_request import VideoStreamRequest  # noqa: E501
from swagger_server.models.video_stream_response import VideoStreamResponse  # noqa: E501
from swagger_server import util


def ai_community_submit_post(body):  # noqa: E501
    """Submit community observations.

    This endpoint allows users to submit images, videos, or descriptions of Mastomys observations. Submissions are reviewed manually or via AI for further analysis.  # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: CommunityObservationResponse
    """
    if connexion.request.is_json:
        body = CommunityObservationRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def ai_detections_post(body):  # noqa: E501
    """Record detected patterns of Mastomys Natalensis populations.

    This endpoint processes uploaded images to detect Mastomys Natalensis populations. It identifies bounding boxes, confidence scores, and other detection metrics.  # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: DetectionPatternResponse
    """
    if connexion.request.is_json:
        body = DetectionPattern.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def ai_explain_post(body):  # noqa: E501
    """Explain AI predictions.

    This endpoint provides explainable AI outputs for predictions made by the system. It offers insights into the factors influencing specific predictions.  # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: ExplainResponse
    """
    if connexion.request.is_json:
        body = ExplainRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def ai_forecast_risk_analysis_post(body):  # noqa: E501
    """Predict outbreak risk for specific regions.

    This endpoint predicts the risk of Lassa fever outbreaks by analyzing population density, historical data, and environmental risk factors in a specified region.  # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: RiskAnalysisResponse
    """
    if connexion.request.is_json:
        body = RiskAnalysisRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def ai_habitats_geospatial_analyze_post(body):  # noqa: E501
    """Perform geospatial habitat analysis.

    This endpoint generates geospatial heatmaps and GeoJSON data for Mastomys habitat suitability. It supports temporal analysis for long-term studies.  # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: GeospatialAnalysisResponse
    """
    if connexion.request.is_json:
        body = GeospatialAnalysisRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def ai_habitats_post(body):  # noqa: E501
    """Analyze satellite and environmental data to identify potential habitats.

    This endpoint analyzes satellite imagery and environmental data to determine habitat suitability for Mastomys Natalensis. It generates suitability scores and identifies key ecological risk factors in the specified region.  # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: HabitatPrediction
    """
    if connexion.request.is_json:
        body = HabitatAnalysisRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def ai_iot_ingest_post():  # noqa: E501
    """Ingest IoT sensor data for real-time monitoring.

    This endpoint processes live IoT sensor data for real-time monitoring of Mastomys habitats. It validates and preprocesses the sensor readings for further analysis.  # noqa: E501


    :rtype: IoTIngestResponse
    """
    return 'do some magic!'


def ai_modeling_post(body):  # noqa: E501
    """Train and evaluate predictive models for ecological analysis.

    This endpoint trains and evaluates predictive models using uploaded datasets for Mastomys Natalensis habitat and population analysis. It supports various model types like LSTM, XGBoost, and Random Forest.  # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: ModelTrainingResponse
    """
    if connexion.request.is_json:
        body = ModelTrainingRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def ai_rag_query_post(body):  # noqa: E501
    """Perform Retrieval-Augmented Generation (RAG) queries.

    This endpoint answers user queries by retrieving relevant information using Retrieval-Augmented Generation (RAG). It combines retrieval capabilities with AI to deliver precise and explainable responses.  # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: RAGQueryResponse
    """
    if connexion.request.is_json:
        body = RAGQueryRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def ai_video_stream_analyze_post(body):  # noqa: E501
    """Analyze live video streams for Mastomys detection.

    This endpoint processes live video streams from sources like drones or stationary cameras. It detects Mastomys populations and generates an annotated video with detection summaries.  # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: VideoStreamResponse
    """
    if connexion.request.is_json:
        body = VideoStreamRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def data_management_open_post(body):  # noqa: E501
    """Open and load datasets for analysis.

    This endpoint allows loading datasets stored externally or on Supabase for further analysis or preprocessing. It supports formats like CSV, GeoJSON, and JSON.  # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: DataManagementOpenResponse
    """
    if connexion.request.is_json:
        body = DataManagementOpenRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def data_management_transform_post(body):  # noqa: E501
    """Transform datasets for compatibility and analysis.

    This endpoint applies transformations like normalization, scaling, or encoding to prepare datasets for analysis or machine learning purposes.  # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: DataManagementTransformResponse
    """
    if connexion.request.is_json:
        body = DataManagementTransformRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'